+++ src/internet/model/tcp-tx-buffer.cc
@@ -148,6 +148,22 @@ TcpTxBuffer::GetSacked() const
     return m_sackedOut;
 }
 
+SequenceNumber32
+TcpTxBuffer::GetLastPacket()
+{
+    auto it = m_sentList.end();
+    it--;
+    TcpTxItem* packet = *it;
+
+    while ((packet->m_sacked || packet->m_lost) && it != m_sentList.begin())
+    {
+        it--;
+        packet = *it;
+    }
+
+    return packet->m_startSeq;
+}
+
 void
 TcpTxBuffer::SetHeadSequence(const SequenceNumber32& seq)
 {
@@ -651,6 +667,27 @@ TcpTxBuffer::IsRetransmittedDataAcked(const SequenceNumber32& ack) const
 }
 
 void
+TcpTxBuffer::GetPacketInfo(SequenceNumber32 ack, TcpTxItem* item)
+{
+    NS_LOG_FUNCTION(this);
+    uint32_t pktSize;
+
+    for (auto it = m_sentList.begin(); it != m_sentList.end(); ++it)
+    {
+        TcpTxItem* packet = *it;
+        Ptr<Packet> p = packet->m_packet;
+        pktSize = p->GetSize();
+
+        // Find out the recent most packet acknowledged
+        if ((packet->m_startSeq == ack && packet->m_sacked) || packet->m_startSeq + pktSize == ack)
+        {
+            *item = *packet;
+            return;
+        }
+    }
+}
+
+void
 TcpTxBuffer::DiscardUpTo(const SequenceNumber32& seq, const Callback<void, TcpTxItem*>& beforeDelCb)
 {
     NS_LOG_FUNCTION(this << seq);
@@ -774,6 +811,11 @@ TcpTxBuffer::Update(const TcpOptionSack::SackList& list, const Callback<void, Tc
     NS_LOG_INFO("Updating scoreboard, got " << list.size() << " blocks to analyze");
 
     uint32_t bytesSacked = 0;
+
+    if (m_sentList.empty())
+    {
+        return false;
+    }
 
     for (auto option_it = list.begin(); option_it != list.end(); ++option_it)
     {
@@ -908,6 +950,107 @@ TcpTxBuffer::UpdateLostCount()
         }
     }
     NS_LOG_INFO("Status after the update: " << *this);
+    ConsistencyCheck();
+}
+
+void
+TcpTxBuffer::RackMarkLossesOnRto(Ptr<TcpRack> rack)
+{
+    NS_LOG_FUNCTION(this);
+
+    if (m_sentList.empty())
+    {
+        return;
+    }
+
+    Time now = Simulator::Now();
+
+    bool firstUnacked = true;
+
+    for (auto it = m_sentList.begin(); it != m_sentList.end(); ++it)
+    {
+        TcpTxItem* item = *it;
+
+        if (item->m_sacked)
+        {
+            continue;
+        }
+
+        bool markLost = false;
+
+        if (firstUnacked)
+        {
+            markLost = true;
+            firstUnacked = false;
+        }
+        else
+        {
+            Time expire = item->m_lastSent + rack->GetRtt() + Seconds(rack->GetReoWnd());
+
+            if (expire <= now)
+            {
+                markLost = true;
+            }
+        }
+
+        if (markLost && !item->m_lost)
+        {
+            item->m_lost = true;
+            m_lostOut += item->m_packet->GetSize();
+
+            if (item->m_retrans)
+            {
+                item->m_retrans = false;
+                m_retrans -= item->m_packet->GetSize();
+            }
+        }
+    }
+
+    ConsistencyCheck();
+}
+
+void
+TcpTxBuffer::DetectRackLoss(Ptr<TcpRack> rack, double* timeout)
+{
+    NS_LOG_FUNCTION(this);
+
+    for (auto it = m_sentList.begin(); it != m_sentList.end(); ++it)
+    {
+        TcpTxItem* item = *it;
+
+        if (item->m_sacked ||
+            !rack->SentAfter(rack->GetXmitTs(),
+                             item->m_lastSent,
+                             rack->GetEndSeq(),
+                             item->m_startSeq.GetValue() + item->m_packet->GetSize()))
+        {
+            continue;
+        }
+
+        double remaining = item->m_lastSent.GetMilliSeconds() + rack->GetRtt().GetMilliSeconds() +
+                           Seconds(rack->GetReoWnd()).GetMilliSeconds() -
+                           Simulator::Now().GetMilliSeconds();
+
+        if (remaining <= 0)
+        {
+            if (!item->m_lost)
+            {
+                item->m_lost = true;
+                m_lostOut += item->m_packet->GetSize();
+            }
+            // Marking the retransmitted packets that are lost again
+            else if (item->m_retrans)
+            {
+                item->m_retrans = false;
+                m_retrans -= item->m_packet->GetSize();
+            }
+        }
+        else
+        {
+            *timeout = std::max(remaining, *timeout);
+        }
+    }
+
     ConsistencyCheck();
 }