+++ RELEASE_NOTES.md
@@ -42,6 +42,8 @@ been tested on Linux. As of this release, the latest known version to work with 
 
 ### New user-visible features
 
+- (internet) !2406 Added support for TCP D-SACK option (RFC 2883).
+
 - (internet) !2405 Added support for TCP FACK (Forward Acknowledgement).
 - (lr-wpan) !2592 Added PHY preamble and RSSI support.
 - (visualizer) Add Lr-Wpan NetDevices support to the Pyviz visualizer.